requirements.txt
json-with-comments
loguru==0.7.3
MaaAgentBinary==1.0.1
MaaFw
numpy>=1.24.0,<2.0.0
pillow>=10.0.0,<11.0.0
python-Levenshtein==0.27.1
RapidFuzz==3.12.2
StrEnum==0.4.15

opencv-python==4.8.1.78
pip~=23.3.1
setuptools~=80.9.0